refactor: re-exported from index.js
