+++ src/context/index.js
@@ -1,4 +1,4 @@
 export { IconProvider } from "./icon-context";
-export { VideoProvider } from "./video-context";
-export { AuthProvider } from "./auth-context";
-export { LikeProvider } from "./like-context";
+export { useVideo, VideoProvider } from "./video-context";
+export { useAuth, AuthProvider } from "./auth-context";
+export { useLike, LikeProvider } from "./like-context";
