+++ src/pages/login-page/LoginPage.jsx
@@ -1,5 +1,5 @@
 import { Link } from "react-router-dom";
-import { useAuth } from "../../context/auth-context";
+import { useAuth } from "../../context";
 import "./login-page.css";
 
 const LoginPage = () => {
