+++ src/pages/signup-page/SignUpPage.jsx
@@ -1,5 +1,5 @@
 import { Link } from "react-router-dom";
-import { useAuth } from "../../context/auth-context";
+import { useAuth } from "../../context";
 import "./signup-page.css";
 
 const SignUpPage = () => {
